streamlining Int/Integer

+++ src/Tasks/LegalProposition/PrintIllegal.hs
@@ -18,7 +18,7 @@ ifUseIllegal useBug notFirstLayer synTree usedStr =
     let nodeNum = treeNodes synTree
     in if not useBug
        then return (normalShow synTree)
-       else frequency [(1, implementIllegal synTree notFirstLayer usedStr),(nodeNum - 1, subTreeIllegal notFirstLayer synTree usedStr)] --  加层数判定
+       else frequency [(1, implementIllegal synTree notFirstLayer usedStr), (fromIntegral nodeNum - 1, subTreeIllegal notFirstLayer synTree usedStr)] --  加层数判定
 
 subTreeIllegal ::Bool -> SynTree Char -> String -> Gen String
 subTreeIllegal notFirstLayer (And a b) usedStr = allocateBugToSubtree notFirstLayer "/\\" a b usedStr
